+++ packages/kit/src/packaging/test/fixtures/typescript/src/lib/Test2.svelte
@@ -1,4 +1,4 @@
 <script lang="ts">
-    import type { Foo } from './foo';
-    export let foo: Foo;
+	import type { Foo } from './foo';
+	export let foo: Foo;
 </script>