+++ packages/kit/types/internal.d.ts
@@ -5,6 +5,7 @@ import {
 	GetSession,
 	Handle,
 	HandleError,
+	KitConfig,
 	Load,
 	RequestEvent,
 	RequestHandler,
@@ -226,6 +227,8 @@ export interface SSREndpoint {
 
 export interface SSRNode {
 	module: SSRComponent;
+	/** index into the `components` array in client-manifest.js */
+	index: number;
 	/** client-side module URL for this component */
 	entry: string;
 	/** external CSS files */
@@ -316,6 +319,8 @@ export interface SSRState {
 export type StrictBody = string | Uint8Array;
 
 export type ValidatedConfig = RecursiveRequired<Config>;
+
+export type ValidatedKitConfig = RecursiveRequired<KitConfig>;
 
 export * from './index';
 export * from './private';