+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,91 @@
 # @sveltejs/kit
 
+## 1.0.0-next.352
+
+### Patch Changes
+
+- fix manifest not updating when adding routes ([#5157](https://github.com/sveltejs/kit/pull/5157))
+
+## 1.0.0-next.351
+
+### Patch Changes
+
+- [fix] URL instance methods now work in `load` ([#5183](https://github.com/sveltejs/kit/pull/5183))
+
+* Upgrade vite-plugin-svelte ([`4e4625ea`](https://github.com/sveltejs/kit/commit/4e4625ea6d9a084bc767ae216704aacd95fe8730))
+
+## 1.0.0-next.350
+
+### Patch Changes
+
+- [breaking] revert removal of `kit.browser.hydrate` ([#5178](https://github.com/sveltejs/kit/pull/5178))
+
+## 1.0.0-next.349
+
+### Patch Changes
+
+- [breaking] remove kit.browser.hydrate config in favor of compilerOptions.hydratable ([#5155](https://github.com/sveltejs/kit/pull/5155))
+
+* [chore] upgrade to Vite 2.9.10 ([#5170](https://github.com/sveltejs/kit/pull/5170))
+
+## 1.0.0-next.348
+
+### Patch Changes
+
+- Disallow access to `__data.json` for standalone endpoints ([#5149](https://github.com/sveltejs/kit/pull/5149))
+
+* [chore] introduce `KitConfig` type ([#5141](https://github.com/sveltejs/kit/pull/5141))
+
+- Prevent incorrect redirect for `__data.json` request with `trailingSlash: 'always'` ([#5149](https://github.com/sveltejs/kit/pull/5149))
+
+## 1.0.0-next.347
+
+### Patch Changes
+
+- Update dependencies ([#5121](https://github.com/sveltejs/kit/pull/5121))
+
+* allow symlinked static assets in dev ([#5089](https://github.com/sveltejs/kit/pull/5089))
+
+## 1.0.0-next.346
+
+### Patch Changes
+
+- [chore] provide Vite config via plugin ([#5073](https://github.com/sveltejs/kit/pull/5073))
+
+* Reload page to recover from HMR errors ([#5108](https://github.com/sveltejs/kit/pull/5108))
+
+- [chore] shared Vite build config ([#5105](https://github.com/sveltejs/kit/pull/5105))
+
+* fix `BodyValidator` handling for nested object literals ([#5118](https://github.com/sveltejs/kit/pull/5118))
+
+- [fix] don't set `credentials` to `undefined` in server-side `fetch` ([#5116](https://github.com/sveltejs/kit/pull/5116))
+
+## 1.0.0-next.345
+
+### Patch Changes
+
+- Prevent component double mounting caused by HMR invalidation ([#4891](https://github.com/sveltejs/kit/pull/4891))
+
+* Only recreate manifest when files inside `config.kit.files.routes` are added or deleted ([#5076](https://github.com/sveltejs/kit/pull/5076))
+
+## 1.0.0-next.344
+
+### Patch Changes
+
+- [fix] svelte-kit sync no longer permanently deletes the types directory ([#5063](https://github.com/sveltejs/kit/pull/5063))
+
+* [chore] trigger sync and other setup from plugin ([#5067](https://github.com/sveltejs/kit/pull/5067))
+
+- Update `engines` to specify that Node 16.7 is required ([#5062](https://github.com/sveltejs/kit/pull/5062))
+
+* only serve `_app/immutable` with immutable cache header, not `_app/version.json` ([#5051](https://github.com/sveltejs/kit/pull/5051))
+
+## 1.0.0-next.343
+
+### Patch Changes
+
+- [chore] return config from server start methods ([#5043](https://github.com/sveltejs/kit/pull/5043))
+
 ## 1.0.0-next.342
 
 ### Patch Changes